+++ src/Avatar/Avatar.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Avatar as ThemeUiAvatar } from 'theme-ui';
 
 export default function Avatar({ ...rest }) {