+++ src/Flex/Flex.stories.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import Flex from './Flex';
 import Box from '../Box/Box';