+++ src/ToggleIcon/ToggleIcon.stories.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import ToggleIcon from './ToggleIcon';
 import { MdFavorite } from 'react-icons/md';
 import Flex from '../Flex/Flex';